requirements.txt
fastapi==0.115.14
uvicorn==0.32.1
python-multipart==0.0.20
sqlalchemy==2.0.36
python-dotenv==1.0.1
pyyaml==6.0.2
opencv-python==4.7.0.72
numpy==1.24.4
scikit-image==0.21.0
scikit-learn==1.3.2
mediapipe==0.10.9
fer==22.5.1
pytest
black
flake8